fix: Move GC to its own namespace
