+++ packages/core/src/category/index.ts
@@ -1,10 +1,7 @@
 import { createSelectSchema } from "drizzle-zod";
-import { categoriesTable, gameCategoriesTable } from "./category.sql";
+import { categoriesTable } from "./category.sql";
 
 export namespace Category {
-    export const CategoriesInfo = createSelectSchema(categoriesTable)
-        .omit({ timeCreated: true, timeDeleted: true, timeUpdated: true })
-
-    export const GameCategoriesInfo = createSelectSchema(gameCategoriesTable)
+    export const Info = createSelectSchema(categoriesTable)
         .omit({ timeCreated: true, timeDeleted: true, timeUpdated: true })
 }
